View와 Controller 클래스 병합과 레시피 수정 기능 상세조회로 변경 완료
